chore: refactor Helpers.cs
