+++ src/Arius.E2E.Tests/Workflows/Steps/AssertConflictBehaviorStep.cs
@@ -16,19 +16,19 @@ internal sealed record AssertConflictBehaviorStep(
 
         Directory.CreateDirectory(state.Fixture.RestoreRoot);
 
-        await RestoreStepSupport.WriteRestoreConflictAsync(
+        await Helpers.WriteRestoreConflictAsync(
             state.Fixture,
             state.Definition,
             ExpectedVersion,
             state.Seed);
 
-        var version = RestoreStepSupport.ResolveVersion(state, Target);
+        var version = Helpers.ResolveVersion(state, Target);
 
-        var result = await RestoreStepSupport.RestoreAsync(state.Fixture, Overwrite, version, cancellationToken);
+        var result = await Helpers.RestoreAsync(state.Fixture, Overwrite, version, cancellationToken);
 
         result.Success.ShouldBeTrue($"{Name}: {result.ErrorMessage}");
 
-        await RestoreStepSupport.AssertRestoreOutcomeAsync(
+        await Helpers.AssertRestoreOutcomeAsync(
             state.Fixture,
             state.Definition,
             ExpectedVersion,
