+++ src/Arius.E2E.Tests/Workflows/Steps/AssertRemoteStateStep.cs
@@ -13,7 +13,7 @@ internal sealed record AssertRemoteStateStep(string Name, RemoteAssertionKind Ki
 {
     public async Task ExecuteAsync(RepresentativeWorkflowState state, CancellationToken cancellationToken)
     {
-        var latest = await WorkflowBlobAssertions.ResolveLatestAsync(state, cancellationToken);
+        var latest = await Helpers.ResolveLatestAsync(state, cancellationToken);
         latest.ShouldNotBeNull($"{Name}: latest snapshot should exist.");
 
         var expectedState = state.CurrentSyntheticRepositoryState
@@ -26,27 +26,27 @@ internal sealed record AssertRemoteStateStep(string Name, RemoteAssertionKind Ki
         switch (Kind)
         {
             case RemoteAssertionKind.InitialArchive:
-                (await WorkflowBlobAssertions.CountBlobsAsync(state.Context.BlobContainer, BlobPaths.Snapshots, cancellationToken))
+                (await Helpers.CountBlobsAsync(state.Context.BlobContainer, BlobPaths.Snapshots, cancellationToken))
                     .ShouldBe(1, $"{Name}: initial archive should create one snapshot.");
                 latest.FileCount.ShouldBe(expectedState.Files.Count, $"{Name}: latest snapshot file count should match the current synthetic dataset state.");
                 break;
 
             case RemoteAssertionKind.IncrementalArchive:
-                (await WorkflowBlobAssertions.CountBlobsAsync(state.Context.BlobContainer, BlobPaths.Snapshots, cancellationToken))
+                (await Helpers.CountBlobsAsync(state.Context.BlobContainer, BlobPaths.Snapshots, cancellationToken))
                     .ShouldBe(2, $"{Name}: incremental archive should create a second snapshot.");
                 latest.FileCount.ShouldBe(expectedState.Files.Count, $"{Name}: latest snapshot file count should match the current synthetic dataset state.");
-                await WorkflowBlobAssertions.AssertLargeDuplicateLookupAsync(state, expectedState, cancellationToken);
-                await WorkflowBlobAssertions.AssertSmallFileTarLookupAsync(state, expectedState, cancellationToken);
+                await Helpers.AssertLargeDuplicateLookupAsync(state, expectedState, cancellationToken);
+                await Helpers.AssertSmallFileTarLookupAsync(state, expectedState, cancellationToken);
                 break;
 
             case RemoteAssertionKind.NoOpArchive:
                 state.PreviousSnapshotVersion.ShouldNotBeNullOrWhiteSpace($"{Name}: previous snapshot version should be available.");
-                var previous = await WorkflowBlobAssertions.ResolveVersionAsync(state, state.PreviousSnapshotVersion, cancellationToken);
+                var previous = await Helpers.ResolveVersionAsync(state, state.PreviousSnapshotVersion, cancellationToken);
                 previous.ShouldNotBeNull($"{Name}: previous snapshot should exist.");
                 latest.RootHash.ShouldBe(previous.RootHash, $"{Name}: no-op archive should preserve the root hash.");
 
-                var chunkCount = await WorkflowBlobAssertions.CountBlobsAsync(state.Context.BlobContainer, BlobPaths.Chunks, cancellationToken);
-                var fileTreeCount = await WorkflowBlobAssertions.CountBlobsAsync(state.Context.BlobContainer, BlobPaths.FileTrees, cancellationToken);
+                var chunkCount = await Helpers.CountBlobsAsync(state.Context.BlobContainer, BlobPaths.Chunks, cancellationToken);
+                var fileTreeCount = await Helpers.CountBlobsAsync(state.Context.BlobContainer, BlobPaths.FileTrees, cancellationToken);
 
                 chunkCount.ShouldBe(
                     state.ChunkBlobCountBeforeNoOpArchive ?? throw new InvalidOperationException($"{Name}: pre-no-op chunk blob count was not captured."),
