+++ src/Arius.E2E.Tests/Workflows/Steps/Helpers.cs
@@ -1,3 +1,4 @@
+using Arius.Core.Features.RestoreCommand;
 using Arius.Core.Shared.ChunkIndex;
 using Arius.Core.Shared.Snapshot;
 using Arius.Core.Shared.Storage;
@@ -6,12 +7,95 @@ using Arius.E2E.Tests.Fixtures;
 
 namespace Arius.E2E.Tests.Workflows.Steps;
 
-internal static class WorkflowBlobAssertions
+internal static class Helpers
 {
     private const string DuplicateLargePathA = "archives/duplicates/binary-a.bin";
     private const string DuplicateLargePathB = "nested/deep/a/b/c/binary-b.bin";
     private const string DuplicateSmallPathA = "nested/deep/a/b/c/d/e/f/copy-b.bin";
     private const string DuplicateSmallPathB = "nested/deep/a/b/c/d/e/f/g/h/copy-c.bin";
+
+    public static Task<RestoreResult> RestoreAsync(E2EFixture fixture, bool overwrite, string? version, CancellationToken cancellationToken)
+    {
+        var options = new RestoreOptions
+        {
+            RootDirectory = fixture.RestoreRoot,
+            Overwrite = overwrite,
+            Version = version,
+        };
+
+        return fixture.CreateRestoreHandler().Handle(new RestoreCommand(options), cancellationToken).AsTask();
+    }
+
+    public static async Task AssertRestoreOutcomeAsync(
+        E2EFixture fixture,
+        SyntheticRepositoryDefinition definition,
+        SyntheticRepositoryVersion expectedVersion,
+        int seed,
+        bool useNoPointers,
+        RestoreResult restoreResult,
+        bool preserveConflictBytes)
+    {
+        if (preserveConflictBytes)
+        {
+            var conflictPath = GetConflictPath(definition, expectedVersion);
+            var restoredPath = Path.Combine(fixture.RestoreRoot, conflictPath.Replace('/', Path.DirectorySeparatorChar));
+            var expectedConflictBytes = CreateConflictBytes(seed, conflictPath);
+
+            restoreResult.FilesSkipped.ShouldBeGreaterThan(0);
+            (await File.ReadAllBytesAsync(restoredPath)).ShouldBe(expectedConflictBytes);
+            return;
+        }
+
+        var expectedRoot = Path.Combine(Path.GetTempPath(), $"arius-expected-{Guid.NewGuid():N}");
+        try
+        {
+            var expected = await SyntheticRepositoryMaterializer.MaterializeAsync(
+                definition,
+                expectedVersion,
+                seed,
+                expectedRoot);
+
+            await SyntheticRepositoryStateAssertions.AssertMatchesDiskTreeAsync(expected, fixture.RestoreRoot, includePointerFiles: false);
+
+            if (!useNoPointers)
+            {
+                foreach (var relativePath in expected.Files.Keys)
+                {
+                    var pointerPath = Path.Combine(
+                        fixture.RestoreRoot,
+                        (relativePath + ".pointer.arius").Replace('/', Path.DirectorySeparatorChar));
+
+                    File.Exists(pointerPath).ShouldBeTrue($"Expected pointer file for {relativePath}");
+                }
+            }
+        }
+        finally
+        {
+            if (Directory.Exists(expectedRoot))
+                Directory.Delete(expectedRoot, recursive: true);
+        }
+    }
+
+    public static async Task WriteRestoreConflictAsync(
+        E2EFixture fixture,
+        SyntheticRepositoryDefinition definition,
+        SyntheticRepositoryVersion expectedVersion,
+        int seed)
+    {
+        var conflictPath = GetConflictPath(definition, expectedVersion);
+        var fullPath = Path.Combine(fixture.RestoreRoot, conflictPath.Replace('/', Path.DirectorySeparatorChar));
+        Directory.CreateDirectory(Path.GetDirectoryName(fullPath)!);
+
+        var conflictBytes = CreateConflictBytes(seed, conflictPath);
+        await File.WriteAllBytesAsync(fullPath, conflictBytes);
+    }
+
+    public static string? ResolveVersion(RepresentativeWorkflowState state, WorkflowRestoreTarget target) =>
+        target switch
+        {
+            WorkflowRestoreTarget.Previous => state.PreviousSnapshotVersion ?? throw new InvalidOperationException("Previous snapshot version is not available."),
+            _ => null,
+        };
 
     public static async Task<int> CountBlobsAsync(IBlobContainerService blobContainer, string prefix, CancellationToken cancellationToken)
     {
@@ -27,9 +111,6 @@ internal static class WorkflowBlobAssertions
 
     public static Task<SnapshotManifest?> ResolveVersionAsync(RepresentativeWorkflowState state, string version, CancellationToken cancellationToken)
         => state.Fixture.Snapshot.ResolveAsync(version, cancellationToken);
-
-    static Task<ShardEntry?> LookupChunkAsync(RepresentativeWorkflowState state, string contentHash, CancellationToken cancellationToken)
-        => state.Fixture.Index.LookupAsync(contentHash, cancellationToken);
 
     public static async Task AssertLargeDuplicateLookupAsync(RepresentativeWorkflowState state, SyntheticRepositoryState expectedState, CancellationToken cancellationToken)
     {
@@ -68,6 +149,26 @@ internal static class WorkflowBlobAssertions
         tarMetadata.Exists.ShouldBeTrue();
         tarMetadata.Metadata.TryGetValue(BlobMetadataKeys.AriusType, out var tarType).ShouldBeTrue();
         tarType.ShouldBe(BlobMetadataKeys.TypeTar);
+    }
+
+    static Task<ShardEntry?> LookupChunkAsync(RepresentativeWorkflowState state, string contentHash, CancellationToken cancellationToken)
+        => state.Fixture.Index.LookupAsync(contentHash, cancellationToken);
+
+    static string GetConflictPath(SyntheticRepositoryDefinition definition, SyntheticRepositoryVersion expectedVersion)
+    {
+        const string v1ChangedPath = "src/module-00/group-00/file-0000.bin";
+
+        if (definition.Files.Any(file => file.Path == v1ChangedPath) && expectedVersion == SyntheticRepositoryVersion.V1)
+            return v1ChangedPath;
+
+        return definition.Files[0].Path;
+    }
+
+    static byte[] CreateConflictBytes(int seed, string path)
+    {
+        var bytes = new byte[1024];
+        new Random(HashCode.Combine(seed, path, "restore-conflict")).NextBytes(bytes);
+        return bytes;
     }
 
     static async Task<string> AssertDuplicateContentHashAsync(RepresentativeWorkflowState state, SyntheticRepositoryState expectedState, string pathA, string pathB, CancellationToken cancellationToken)
